Adapt game to new data structure
