+++ src/index.js
@@ -14,7 +14,7 @@ const assets = require('./utils/assets')
 const choo = require('choo')
 const hotModuleReplacement = require('./utils/hot-module-replacement')
 const ideModel = require('./model')
-const gameModel = require('./game')
+const gameModel = require('./game/model')
 
 const app = choo()
 
@@ -34,10 +34,10 @@ window.startApp = () => {
     WATER_BLOCK: '../assets/img/water-block.png',
     STONE_BLOCK: '../assets/img/stone-block.png',
     GEM: '../assets/img/gem-blue.png',
-    [`ROBOT_${ORIENTATION.FRONT}`]: '../assets/img/robot-front.png',
-    [`ROBOT_${ORIENTATION.LEFT}`]: '../assets/img/robot-left.png',
-    [`ROBOT_${ORIENTATION.BACK}`]: '../assets/img/robot-back.png',
-    [`ROBOT_${ORIENTATION.RIGHT}`]: '../assets/img/robot-right.png'
+    ROBOT_FRONT: '../assets/img/robot-front.png',
+    ROBOT_LEFT: '../assets/img/robot-left.png',
+    ROBOT_BACK: '../assets/img/robot-back.png',
+    ROBOT_RIGHT: '../assets/img/robot-right.png'
   }).then(() => {
     document.body.innerHTML = ''
     document.body.appendChild(app.start())
